feat: implement alert suggestion parsing and PR comment creation functionality
